third_party/WebKit/LayoutTests/fast/dom/shadow/style-of-distributed-node-expected.txt

CONSOLE WARNING: line 27: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests to ensure that styles of distributed nodes are different if their parent styles are different.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".





PASS getColorProperty("child-a") is "rgb(255, 0, 0)"
PASS getColorProperty("child-b") is "rgb(0, 0, 255)"




PASS successfullyParsed is true

TEST COMPLETE
